+++ tests/test_unresolved_package_handler.py
@@ -21,7 +21,7 @@ import pytest
 
 from pathlib import Path
 
-from unresolved_package_handler import unresolved_package_handler
+from thoth.unresolved_package_handler.unresolved_package_handler import unresolved_package_handler
 
 
 class TestProducer:
@@ -30,6 +30,7 @@ class TestProducer:
     def test_unresolved_package_handler(self) -> None:
         """Test extraction of inputs for Kafka."""
         file_test_path = Path().cwd().joinpath("tests", "adviser-04ab56d6.json")
+        print(file_test_path)
         unresolved_packages, package_version, sources, solver = unresolved_package_handler(
             file_test_path=file_test_path
         )
